feat: 위키 에디터 메뉴, 이름 헷갈릴까봐 이름 변경, 버튼컴포넌트 분리
