+++ src/components/WikiEditor.tsx
@@ -3,7 +3,7 @@ import StarterKit from "@tiptap/starter-kit";
 import { Color } from "@tiptap/extension-color";
 import ListItem from "@tiptap/extension-list-item";
 import TextStyle from "@tiptap/extension-text-style";
-import EditorMenu from "./EditorMenu";
+import EditorMenu from "./WikiEditorMenu";
 
 export default function WikiEditor() {
   const editor = useEditor({
